Stringify node ids before creating edge ids
Fixes #47

This change fixes a bug specific to undirected graphs.

Prior to this commit we did not consistently stringify node ids before
generating an edge id. We sort the node ids when generating the edge id
for undirected graphs to ensure that updates and queries always use the
same id. This sort yielded inconsistent results for numbers (9 < 10, but
"9" > "10").