Separate index.js from CJS_FILES in Makefile:
It's inconvenient to have index.js be in CJS_FILES because:

CJS_FILES are the target deps for `make convert`
And `convert` is a target dep for for `index.js`
So they both want each other to updated and Makefile complains about
circular dependencies.

Errata:

CJS_FILES (originally SRC_FILES) contained index.js AND index.js was
also specified next to SRC_FILES at the same time (so being specified
twice), so even originally it was a bit weird.